+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.3.0 (Unreleased)
+
+- Type of `WeatherInfo` is changed to make it move away from `String`
+  based result.
+- Fix issues while parsing weather data for VOGO and KYKM stations.
+
 # v0.2.0
 
 * Add blocking version of get_weather function.